+++ src/CommandPaletteInternal.h
@@ -43,7 +43,6 @@ struct CommandPaletteWnd : WindowBase {
     StrVecCP favorites;
     ListBox* listBox = nullptr;
     Static* staticInfo = nullptr;
-    Checkbox* checkboxPre36 = nullptr;
 
     StrVec filterWords;
     Vec<u8> highlighted;
@@ -79,7 +78,6 @@ struct CommandPaletteWnd : WindowBase {
     void SwitchToFavorites();
     void OnSelectionChange();
     void OnListDoubleClick();
-    void OnPre36CheckboxChanged();
     void DrawListBoxItem(ListBox::DrawItemEvent* ev);
 };