+++ src/SumatraPDF.cpp
@@ -9811,8 +9811,8 @@ static LRESULT FrameOnCommand(MainWindow* win, HWND hwnd, UINT msg, WPARAM wp, L
 
         case CmdNextTabSmart:
         case CmdPrevTabSmart: {
-            if (gGlobalPrefs->ctrlTabPre36Behavior) {
-                // pre-3.6 behavior: switch tabs immediately, in tab-strip order
+            if (gGlobalPrefs->ctrlTabSimple) {
+                // simple (pre-3.6) behavior: switch tabs immediately, in tab-strip order
                 TabsOnCtrlTab(win, cmdId == CmdPrevTabSmart);
                 break;
             }